Adapt test file names as well and add Traits
